% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpetual-package.R
\docType{package}
\name{perpetual-package}
\alias{perpetual-package}
\alias{class-package}
\title{perpetual: PerpetualBooster}
\description{
A self-generalizing gradient boosting machine that doesn't need hyperparameter optimization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://perpetual-ml.github.io/perpetual/r/}
}

}
\author{
\strong{Maintainer}: Mutlu Simsek \email{msimsek@perpetual-ml.com}

Authors:
\itemize{
  \item Serkan Korkmaz \email{serkor1@duck.com}
  \item Pieter Pel \email{pelpieter@gmail.com}
}

}
\keyword{package}
