% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{perpetual_calibrate}
\alias{perpetual_calibrate}
\title{Calibrate a PerpetualBooster model}
\usage{
perpetual_calibrate(model, x, y, x_cal, y_cal, alpha)
}
\arguments{
\item{model}{A \code{PerpetualBooster} object.}

\item{x}{Validation features.}

\item{y}{Validation targets.}

\item{x_cal}{Calibration features.}

\item{y_cal}{Calibration targets.}

\item{alpha}{Calibration parameter.}
}
\description{
Calibrate a PerpetualBooster model
}
