% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{perpetual_load}
\alias{perpetual_load}
\title{Load a PerpetualBooster model}
\usage{
perpetual_load(path)
}
\arguments{
\item{path}{String, path to the saved model.}
}
\value{
A \code{PerpetualBooster} object.
}
\description{
Load a PerpetualBooster model
}
