% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{perpetual_n_trees}
\alias{perpetual_n_trees}
\title{Get the number of trees in the model}
\usage{
perpetual_n_trees(model)
}
\arguments{
\item{model}{A \code{PerpetualBooster} object.}
}
\value{
Integer, number of trees.
}
\description{
Get the number of trees in the model
}
