% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.PerpetualBooster}
\alias{print.PerpetualBooster}
\title{Print PerpetualBooster}
\usage{
\method{print}{PerpetualBooster}(x, ...)
}
\arguments{
\item{x}{A \code{PerpetualBooster} object.}

\item{...}{Additional arguments.}
}
\description{
Print PerpetualBooster
}
